requirements.txt
gradio
openai
markitdown[pdf]
python-dotenv